requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
websocket-client>=1.6.0
prompt-toolkit>=3.0.39
rich>=13.5.0
pydantic>=2.4.0
cryptography>=41.0.0
PyNaCl>=1.5.0
